Can read from DB but not yet write

+++ src/main/java/com/humanbooster/zobi/dao/AnimalDaoInterface.java
@@ -11,7 +11,7 @@ import com.humanbooster.zobi.business.Animal;
  * @author humanbooster
  *
  */
-public interface AnimalDaoInterface extends Dao<Integer, Animal> {
+public interface AnimalDaoInterface extends Dao<Long, Animal> {
 
 	
 	/**
