+++ src/main/java/com/humanbooster/zobi/dao/SpeciesDaoInterface.java
@@ -9,6 +9,6 @@ import com.humanbooster.zobi.business.Species;
  * @author humanbooster
  *
  */
-public interface SpeciesDaoInterface extends Dao<Integer, Species> {
+public interface SpeciesDaoInterface extends Dao<Long, Species> {
 	
 }
